Huseyn

+++ src/main/java/az/example/eventsapp/response/EventCriteria.java
@@ -1,6 +1,7 @@
 package az.example.eventsapp.response;
 
 import lombok.Data;
+import org.springframework.format.annotation.DateTimeFormat;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
@@ -10,8 +11,8 @@ public class EventCriteria {
     private String name;
     private LocalDateTime startDateFrom;
     private LocalDateTime startDateTo;
-    private BigDecimal minPrice;
-    private BigDecimal maxPrice;
+    private Double minPrice;
+    private Double maxPrice;
     private String venueName;
     private String sortBy;
 }
